+++ src/merkle/smt/mod.rs
@@ -56,6 +56,9 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
     /// The default value used to compute the hash of empty leaves
     const EMPTY_VALUE: Self::Value;
 
+    /// The root of the empty tree with provided DEPTH
+    const EMPTY_ROOT: RpoDigest;
+
     // PROVIDED METHODS
     // ---------------------------------------------------------------------------------------------